Добавлены тесты для генерации тестирующего класса
